requirements.txt
discord.py
python-dotenv
gspread
pandas